chore: rename project, configuration and linting
