+++ pyproject.toml
@@ -3,14 +3,14 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "census-viz"
+name = "abheile"
 version = "0.1.0"
-description = "Visualization tool for US Census data"
+description = "Spatial visualization tool for demographic and economic data"
 readme = "README.md"
 requires-python = ">=3.13"
 dependencies = [
-    "deltalake>=0.22.3",
-    "polars>=1.18.0",
+    "deltalake~=0.25",
+    "polars~=1.28",
     "pydantic>=2.10.4",
     "httpx>=0.27.0",  # Modern async HTTP client
     "pydantic-settings>=2.1.0",  # Settings management
@@ -29,7 +29,7 @@ dev = [
     "ruff>=0.3.0",
 ]
 notebook = [
-    "marimo>=0.10.9",
+    "marimo~=0.13",
 ]
 
 [tool.hatch.build.targets.wheel]
@@ -39,10 +39,16 @@ packages = ["src/census_viz"]
 line-length = 88
 target-version = "py313"
 
+[tool.ruff.lint]
+select = ["A", "ANN", "E", "F", "I", "SIM", "UP"]
+
 [tool.pytest.ini_options]
 asyncio_mode = "auto"
 testpaths = ["tests"]
 python_files = ["test_*.py"]
+
+[tool.marimo.runtime]
+output_max_bytes = 25_000_000
 
 [project.scripts]
 census-collect = "census_viz.cli:app"
