--- src/census_viz/__init__.py
@@ -1,13 +1,0 @@
-"""Census visualization package"""
-
-from census_viz.models import CensusResponse, GeographicArea, PopulationData
-from census_viz.client.census import CensusClient
-from census_viz.client.tiger import TigerClient
-
-__all__ = [
-    "CensusClient",
-    "TigerClient",
-    "CensusResponse",
-    "GeographicArea",
-    "PopulationData",
-]
