--- src/census_viz/client/__init__.py
@@ -1,5 +1,0 @@
-from census_viz.client.census import CensusClient
-from census_viz.client.tiger import TigerClient
-
-
-__all__ = ["CensusClient", "TigerClient"]
