+++ tests/client/test_census.py
@@ -1,7 +1,9 @@
-import pytest
-from census_viz.client.census import CensusClient
-from census_viz.config import settings
 import os
+
+import pytest
+
+from abheile.client.census import CensusClient
+from abheile.config import settings
 
 
 @pytest.fixture
